requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-qdrant
langgraph
langsmith
streamlit
qdrant-client
pyowm
pypdf
python-dotenv
pytest